추가: request, file crud 테스트 코드 수정: repository 파일 폴더 정리

+++ src/main/java/com/example/demo/repository/DeceasedRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DeceasedRepository extends JpaRepository<Deceased, Integer> {
+public interface DeceasedRepository extends JpaRepository<Deceased, Long> {
 
 }
